like_handler/requirements.txt
fastapi==0.88.0
python-dotenv==0.21.0
requests==2.28.1
uvicorn==0.20.0
psycopg2==2.9.5
firebase-admin
numpy==1.24.0